+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2025-04-24
+
+### Added
+
+- Benchmark comparing  `ThreadMap` and `ThreadLocal`.
+
+### Changed
+
+- Fixed the benchmark comparing `ThreadMap` and `ThreadMapX` from the previous version, as it was properly constructed.
+- Benchmarks section of library documentation.
+
 ## [1.0.2] - 2025-04-24
 
 This release is backward compatible with the previous one.